add frame compare teste
